1.5

+++ tasks/makovskiy_i_gauss_filter_vert/seq/src/ops_seq.cpp
@@ -1,12 +1,7 @@
 #include "makovskiy_i_gauss_filter_vert/seq/include/ops_seq.hpp"
 
-#include <algorithm>
 #include <array>
 #include <cstddef>
-#include <tuple>
-#include <vector>
-
-#include "makovskiy_i_gauss_filter_vert/common/include/common.hpp"
 
 namespace makovskiy_i_gauss_filter_vert {
 
@@ -39,7 +34,8 @@ bool GaussFilterSEQ::RunImpl() {
       int sum = 0;
       for (int k_row = -1; k_row <= 1; ++k_row) {
         for (int k_col = -1; k_col <= 1; ++k_col) {
-          sum += GetPixel(input, col + k_col, row + k_row, width, height) * kernel[((k_row + 1) * 3) + (k_col + 1)];
+          sum += GetPixel(input, col + k_col, row + k_row, width, height) *
+                 kernel[static_cast<size_t>((k_row + 1) * 3 + (k_col + 1))];
         }
       }
       output[(static_cast<size_t>(row) * width) + col] = sum / kernel_sum;
